+++ mike/commands.py
@@ -14,6 +14,10 @@ from .versions import Versions
 versions_file = 'versions.json'
 
 
+def _format_deploy_prefix(deploy_prefix):
+    return ' in {}'.format(deploy_prefix) if deploy_prefix else ''
+
+
 def _redirect_template(user_template=None):
     f = (open(user_template, 'rb') if user_template else
          resource_stream(__name__, 'templates/redirect.html'))
@@ -32,18 +36,18 @@ def _add_redirect_to_commit(commit, template, src, dst,
         commit.add_file(git_utils.FileInfo(src, template.render(href=href)))
 
 
-def list_versions(branch='gh-pages', prefix=''):
+def list_versions(branch='gh-pages', deploy_prefix=''):
     try:
         return Versions.loads(git_utils.read_file(
-            branch, os.path.join(prefix, versions_file),
+            branch, os.path.join(deploy_prefix, versions_file),
             universal_newlines=True
         ))
     except git_utils.GitError:
         return Versions()
 
 
-def versions_to_file_info(versions, prefix=''):
-    return git_utils.FileInfo(os.path.join(prefix, versions_file),
+def versions_to_file_info(versions, deploy_prefix=''):
+    return git_utils.FileInfo(os.path.join(deploy_prefix, versions_file),
                               versions.dumps())
 
 
@@ -54,24 +58,24 @@ def make_nojekyll():
 @contextmanager
 def deploy(cfg, version, title=None, aliases=[], update_aliases=False,
            redirect=True, template=None, *, branch='gh-pages', message=None,
-           prefix=''):
+           deploy_prefix=''):
     if message is None:
         message = (
-            'Deployed {rev} to {doc_version}{prefix} with MkDocs ' +
+            'Deployed {rev} to {doc_version}{deploy_prefix} with MkDocs ' +
             '{mkdocs_version} and mike {mike_version}'
         ).format(
             rev=git_utils.get_latest_commit('HEAD', short=True),
             doc_version=version,
-            prefix=' in {}'.format(prefix) if prefix else '',
+            deploy_prefix=_format_deploy_prefix(deploy_prefix),
             mkdocs_version=mkdocs_utils.version(),
             mike_version=app_version
         )
 
-    all_versions = list_versions(branch, prefix)
+    all_versions = list_versions(branch, deploy_prefix)
     info = all_versions.add(version, title, aliases, update_aliases)
     version_str = str(info.version)
-    destdir = os.path.join(prefix, version_str)
-    alias_destdirs = [os.path.join(prefix, i) for i in info.aliases]
+    destdir = os.path.join(deploy_prefix, version_str)
+    alias_destdirs = [os.path.join(deploy_prefix, i) for i in info.aliases]
 
     # Let the caller perform the build.
     yield
@@ -95,32 +99,32 @@ def deploy(cfg, version, title=None, aliases=[], update_aliases=False,
                 else:
                     commit.add_file(alias_file)
 
-        commit.add_file(versions_to_file_info(all_versions, prefix))
+        commit.add_file(versions_to_file_info(all_versions, deploy_prefix))
         commit.add_file(make_nojekyll())
 
 
 def delete(versions=None, all=False, *, branch='gh-pages', message=None,
-           prefix=''):
+           deploy_prefix=''):
     if not all and versions is None:
         raise ValueError('specify `version` or `all`')
 
     if message is None:
         message = (
-            'Removed {doc_version}{prefix} with mike {mike_version}'
+            'Removed {doc_version}{deploy_prefix} with mike {mike_version}'
         ).format(
             doc_version='everything' if all else ', '.join(versions),
-            prefix=' in {}'.format(prefix) if prefix else '',
+            deploy_prefix=_format_deploy_prefix(deploy_prefix),
             mike_version=app_version
         )
 
     with git_utils.Commit(branch, message) as commit:
         if all:
-            if prefix:
-                commit.delete_files([prefix])
+            if deploy_prefix:
+                commit.delete_files([deploy_prefix])
             else:
                 commit.delete_files('*')
         else:
-            all_versions = list_versions(branch, prefix)
+            all_versions = list_versions(branch, deploy_prefix)
             try:
                 removed = all_versions.difference_update(versions)
             except KeyError as e:
@@ -128,18 +132,18 @@ def delete(versions=None, all=False, *, branch='gh-pages', message=None,
 
             for i in removed:
                 if isinstance(i, str):
-                    commit.delete_files([os.path.join(prefix, i)])
+                    commit.delete_files([os.path.join(deploy_prefix, i)])
                 else:
                     commit.delete_files(
-                        [os.path.join(prefix, str(i.version))] +
-                        [os.path.join(prefix, j) for j in i.aliases]
+                        [os.path.join(deploy_prefix, str(i.version))] +
+                        [os.path.join(deploy_prefix, j) for j in i.aliases]
                     )
-            commit.add_file(versions_to_file_info(all_versions, prefix))
+            commit.add_file(versions_to_file_info(all_versions, deploy_prefix))
 
 
 def alias(cfg, version, aliases, update_aliases=False, redirect=True,
-          template=None, *, branch='gh-pages', message=None, prefix=''):
-    all_versions = list_versions(branch, prefix)
+          template=None, *, branch='gh-pages', message=None, deploy_prefix=''):
+    all_versions = list_versions(branch, deploy_prefix)
     try:
         real_version = all_versions.find(version, strict=True)[0]
     except KeyError as e:
@@ -147,18 +151,18 @@ def alias(cfg, version, aliases, update_aliases=False, redirect=True,
 
     if message is None:
         message = (
-            'Copied {doc_version} to {aliases}{prefix} with mike ' +
+            'Copied {doc_version} to {aliases}{deploy_prefix} with mike ' +
             '{mike_version}'
         ).format(
             doc_version=real_version,
             aliases=', '.join(aliases),
-            prefix=' in {}'.format(prefix) if prefix else '',
+            deploy_prefix=_format_deploy_prefix(deploy_prefix),
             mike_version=app_version
         )
 
     new_aliases = all_versions.update(real_version, aliases=aliases,
                                       update_aliases=update_aliases)
-    destdirs = [os.path.join(prefix, i) for i in new_aliases]
+    destdirs = [os.path.join(deploy_prefix, i) for i in new_aliases]
 
     if redirect and destdirs:
         t = _redirect_template(template)
@@ -166,7 +170,7 @@ def alias(cfg, version, aliases, update_aliases=False, redirect=True,
     with git_utils.Commit(branch, message) as commit:
         commit.delete_files(destdirs)
 
-        canonical_dir = os.path.join(prefix, str(real_version))
+        canonical_dir = os.path.join(deploy_prefix, str(real_version))
         for canonical_file in git_utils.walk_files(branch, canonical_dir):
             for d in destdirs:
                 alias_file = canonical_file.copy(d, canonical_dir)
@@ -177,51 +181,53 @@ def alias(cfg, version, aliases, update_aliases=False, redirect=True,
                     )
                 else:
                     commit.add_file(alias_file)
-        commit.add_file(versions_to_file_info(all_versions, prefix))
+        commit.add_file(versions_to_file_info(all_versions, deploy_prefix))
 
 
-def retitle(version, title, *, branch='gh-pages', message=None, prefix=''):
+def retitle(version, title, *, branch='gh-pages', message=None,
+            deploy_prefix=''):
     if message is None:
         message = (
-            'Set title of {doc_version} to {title}{prefix} with mike ' +
+            'Set title of {doc_version} to {title}{deploy_prefix} with mike ' +
             '{mike_version}'
         ).format(
             doc_version=version,
             title=title,
-            prefix=' in {}'.format(prefix) if prefix else '',
+            deploy_prefix=_format_deploy_prefix(deploy_prefix),
             mike_version=app_version
         )
 
-    all_versions = list_versions(branch, prefix)
+    all_versions = list_versions(branch, deploy_prefix)
     try:
         all_versions.update(version, title)
     except KeyError:
         raise ValueError('version {!r} does not exist'.format(version))
 
     with git_utils.Commit(branch, message) as commit:
-        commit.add_file(versions_to_file_info(all_versions, prefix))
+        commit.add_file(versions_to_file_info(all_versions, deploy_prefix))
 
 
 def set_default(version, template=None, *, branch='gh-pages', message=None,
-                prefix=''):
+                deploy_prefix=''):
     if message is None:
         message = (
-            'Set default version to {doc_version}{prefix} with mike ' +
+            'Set default version to {doc_version}{deploy_prefix} with mike ' +
             '{mike_version}'
         ).format(
             doc_version=version,
-            prefix=' in {}'.format(prefix) if prefix else '',
+            deploy_prefix=_format_deploy_prefix(deploy_prefix),
             mike_version=app_version
         )
 
-    all_versions = list_versions(branch, prefix)
+    all_versions = list_versions(branch, deploy_prefix)
     if not all_versions.find(version):
         raise ValueError('version {!r} does not exist'.format(version))
 
     t = _redirect_template(template)
     with git_utils.Commit(branch, message) as commit:
         commit.add_file(git_utils.FileInfo(
-            os.path.join(prefix, 'index.html'), t.render(href=version + '/')
+            os.path.join(deploy_prefix, 'index.html'),
+            t.render(href=version + '/')
         ))