+++ test/integration/test_delete.py
@@ -7,10 +7,10 @@ from mike import git_utils
 
 
 class DeleteTestCase(unittest.TestCase):
-    def _deploy(self, branch=None, versions=['1.0', '2.0'], prefix=''):
+    def _deploy(self, branch=None, versions=['1.0', '2.0'], deploy_prefix=''):
         extra_args = ['-b', branch] if branch else []
-        if prefix:
-            extra_args.extend(['--prefix', prefix])
+        if deploy_prefix:
+            extra_args.extend(['--deploy-prefix', deploy_prefix])
         for i in versions:
             assertPopen(['mike', 'deploy', i] + extra_args)
 
@@ -81,9 +81,9 @@ class TestDelete(DeleteTestCase):
         check_call_silent(['git', 'checkout', 'gh-pages'])
         self._test_delete('commit message')
 
-    def test_prefix(self):
-        self._deploy(prefix='prefix')
-        assertPopen(['mike', 'delete', '1.0', '--prefix', 'prefix'])
+    def test_deploy_prefix(self):
+        self._deploy(deploy_prefix='prefix')
+        assertPopen(['mike', 'delete', '1.0', '--deploy-prefix', 'prefix'])
         check_call_silent(['git', 'checkout', 'gh-pages'])
         self._test_delete(directory='prefix')